==> notebooks/3.0-models-predict.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import math
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import normalize
from sklearn.metrics import confusion_matrix, f1_score
from sklearn.tree import DecisionTreeClassifier
from sklearn.model_selection import RandomizedSearchCV
from sklearn.ensemble import RandomForestClassifier
from sklearn.calibration import CalibratedClassifierCV
#from xgboost import XGBClassifier
from sklearn.tree import DecisionTreeRegressor
from sklearn.ensemble import RandomForestRegressor
#from xgboost import XGBRegressor
import pickle
import joblib
#from tqdm import tqdm
from sklearn.metrics import mean_squared_error
from sklearn.ensemble import AdaBoostRegressor
#import lightgbm as lgb
import os

path = r"C:\Users\imate\Documents\24.9.Notebooks_training\Energy-predictor\reports"
path2=r'C:\Users\imate\Documents\24.9.Notebooks_training\Energy-predictor\data\2_processed'
path3=r'C:\Users\imate\Documents\24.9.Notebooks_training\Energy-predictor\models'


X_train_f = os.path.join(path2, "2_test.ftr")
X_test= pd.read_feather(X_train_f)

- DT

dt_reg=joblib.load('decision_tree_reg.sav')
predictions=pd.DataFrame()
predictions['row_id']=X_test['row_id']
X_test.drop(['row_id'], axis=1, inplace=True)
predictions['log_meter_reading']= dt_reg.predict(X_test)
predictions['meter_reading']= np.expm1(predictions['log_meter_reading'])
predictions.drop(['log_meter_reading'], axis=1,inplace=True)

output = os.path.join(path, "Decision_tree_predictions.csv")
predictions.to_csv(output, index=False)

- LGB GBDT

model = os.path.join(path3, "lgb_reg.sav")

gbdt_reg=joblib.load(model)

predictions['log_meter_reading']= gbdt_reg.predict(X_test)
predictions.drop(['meter_reading'], axis=1,inplace=True)
predictions['meter_reading']= np.expm1(predictions['log_meter_reading'])
predictions.drop(['log_meter_reading'], axis=1,inplace=True)

output = os.path.join(path, "LGB_GBDT_predictions.csv")
predictions.to_csv(output, index=False)

- RF

rf_reg=joblib.load('rf_reg.sav')
predictions['log_meter_reading']= rf_reg.predict(X_test)
predictions.drop(['meter_reading'], axis=1,inplace=True)
predictions['meter_reading']= np.expm1(predictions['log_meter_reading'])
predictions.drop(['log_meter_reading'], axis=1,inplace=True)

output = os.path.join(path, "RF_predictions.csv")
predictions.to_csv(output, index=False)

- CatBoost GBDT

cb_reg=joblib.load('cb_reg.sav')
predictions['log_meter_reading']= cb_reg.predict(X_test)
predictions['meter_reading']= np.expm1(predictions['log_meter_reading'])
predictions.drop(['log_meter_reading'], axis=1,inplace=True)

output = os.path.join(path, "CB_predictions.csv")
predictions.to_csv(output, index=False)
